move state persistence interface to core interfaces
