+++ src/qubx/utils/runner/factory.py
@@ -7,11 +7,10 @@ import os
 from typing import Any, Optional
 
 from qubx import logger
-from qubx.core.interfaces import IAccountViewer, IMetricEmitter, IStrategyNotifier, ITradeDataExport
+from qubx.core.interfaces import IAccountViewer, IMetricEmitter, IStatePersistence, IStrategyNotifier, ITradeDataExport
 from qubx.data.composite import CompositeReader
 from qubx.data.readers import DataReader
 from qubx.emitters.composite import CompositeMetricEmitter
-from qubx.state.interfaces import IStatePersistence
 from qubx.utils.misc import class_import
 from qubx.utils.runner.configs import (
     EmissionConfig,
